cleaned up the code ;)
