+++ .eslintrc.js
@@ -3,10 +3,11 @@ module.exports = {
     commonjs: true,
     es2021: true,
     node: true,
+    jest: true,
   },
   extends: ['standard', 'prettier'],
   parserOptions: {
     ecmaVersion: 12,
   },
   rules: {},
-}
+};
