+++ routes/api/contacts.js
@@ -1,14 +1,7 @@
 const express = require('express');
 const ctrlWrapper = require('../../middelwares/ctrlWrapper');
 const auth = require('../../middelwares/auth');
-const {
-  getContacts,
-  getContact,
-  addNewContact,
-  changeContact,
-  patchContact,
-  deleteContact,
-} = require('../../controllers/contactsController');
+const ctrl = require('../../controllers/contacts');
 
 const {
   addContactValidation,
@@ -17,11 +10,19 @@ const {
 
 const router = express.Router();
 
-router.get('/', auth, ctrlWrapper(getContacts));
-router.get('/:contactId', ctrlWrapper(getContact));
-router.post('/', auth, addContactValidation, ctrlWrapper(addNewContact));
-router.put('/:contactId', addContactValidation, ctrlWrapper(changeContact));
-router.patch('/:contactId', patchContactValidation, ctrlWrapper(patchContact));
-router.delete('/:contactId', deleteContact);
+router.get('/', auth, ctrlWrapper(ctrl.getContacts));
+router.get('/:contactId', ctrlWrapper(ctrl.getContact));
+router.post('/', auth, addContactValidation, ctrlWrapper(ctrl.addNewContact));
+router.put(
+  '/:contactId',
+  addContactValidation,
+  ctrlWrapper(ctrl.changeContact)
+);
+router.patch(
+  '/:contactId',
+  patchContactValidation,
+  ctrlWrapper(ctrl.patchContact)
+);
+router.delete('/:contactId', ctrl.deleteContact);
 
 module.exports = router;
